Upgrade rustc & pin rocket version
It's happened to me a couple times now, I go to compile aardwolf and it
no longer works because rocket has put out a new patch version and now
rocket_codegen will no longer compile under our pinned rustc version.

So this PR upgrades us to the latest rocket, and the necessary rustc for
it.

But, it also pins the rocket version (using `rocket = "=0.3.5"` instead
of `rocket = "0.3.5"`) so that hopefully we won't have any surprise
breakages due to rocket putting out a new patch verison

It does mean we have to be vigilant about upgrading rocket when we can,
however.