fixed some typos and links on readme (#125)
Not sure if what I did to the link for the license in line 83 is right. There was a square bracket after the link for the svg file, and it was looking strange on output on the readme file here on github. I took out that square bracket but I'm not sure if that was the right thing to do.